// src/auth/signin-and-signup-response-payload.interface.ts
export interface SignInAndSignUpResponsePyaload {
  firstName: string;
  lastName: string;
  displayName: string;
  email: string;
  userType: string;
  balance: number;
  rib: string;
  address: string;
  phoneNumber: number;
  city: string;
  zipCode: string;
  accessToken: string;
  refreshToken: string;
}